requirements.txt
fastapi[all]
transformers
uvicorn
torch
pillow
accelerate
timm